refactor: move operations._submitter to _submitter
