+++ .cspell.json
@@ -149,7 +149,8 @@
     "vcrpy",
     "uionly",
     "llmops",
-    "Abhishek"
+    "Abhishek",
+	"coreclr"
   ],
   "flagWords": [
     "Prompt Flow"
