+++ src/promptflow/promptflow/_sdk/operations/_flow_operations.py
@@ -16,6 +16,7 @@ import yaml
 
 from promptflow._sdk._constants import CHAT_HISTORY, DEFAULT_ENCODING, LOCAL_MGMT_DB_PATH
 from promptflow._sdk._load_functions import load_flow
+from promptflow._sdk._submitter import TestSubmitter
 from promptflow._sdk._utils import (
     _get_additional_includes,
     _merge_local_code_and_additional_includes,
@@ -27,8 +28,6 @@ from promptflow._sdk._utils import (
     parse_variant,
 )
 from promptflow._sdk.entities._validation import ValidationResult
-from promptflow._sdk.operations._run_submitter import remove_additional_includes, variant_overwrite_context
-from promptflow._sdk.operations._test_submitter import TestSubmitter
 from promptflow._telemetry.activity import ActivityType, monitor_operation
 from promptflow._telemetry.telemetry import TelemetryMixin
 from promptflow._utils.context_utils import _change_working_dir
@@ -128,28 +127,54 @@ class FlowOperations(TelemetryMixin):
         inputs = inputs or {}
         flow = load_flow(flow)
         flow.context.variant = variant
-        with TestSubmitter(flow=flow, flow_context=flow.context, client=self._client).init() as submitter:
-            is_chat_flow, chat_history_input_name, _ = self._is_chat_flow(submitter.dataplane_flow)
-            flow_inputs, dependency_nodes_outputs = submitter.resolve_data(
-                node_name=node, inputs=inputs, chat_history_name=chat_history_input_name
-            )
+        if flow.language == "python":
+            with TestSubmitter(flow=flow, flow_context=flow.context, client=self._client).init() as submitter:
+                is_chat_flow, chat_history_input_name, _ = self._is_chat_flow(submitter.dataplane_flow)
+                flow_inputs, dependency_nodes_outputs = submitter.resolve_data(
+                    node_name=node, inputs=inputs, chat_history_name=chat_history_input_name
+                )
 
-            if node:
-                return submitter.node_test(
-                    node_name=node,
-                    flow_inputs=flow_inputs,
-                    dependency_nodes_outputs=dependency_nodes_outputs,
-                    environment_variables=environment_variables,
-                    stream=True,
+                if node:
+                    return submitter.node_test(
+                        node_name=node,
+                        flow_inputs=flow_inputs,
+                        dependency_nodes_outputs=dependency_nodes_outputs,
+                        environment_variables=environment_variables,
+                        stream=True,
+                    )
+                else:
+                    return submitter.flow_test(
+                        inputs=flow_inputs,
+                        environment_variables=environment_variables,
+                        stream_log=stream_log,
+                        stream_output=stream_output,
+                        allow_generator_output=allow_generator_output and is_chat_flow,
+                    )
+        elif flow.language == "csharp":
+            from promptflow._sdk._submitter import CSharpTestSubmitter
+
+            with CSharpTestSubmitter(flow=flow, flow_context=flow.context, client=self._client).init() as submitter:
+                is_chat_flow, chat_history_input_name, _ = self._is_chat_flow(submitter.dataplane_flow)
+                flow_inputs, dependency_nodes_outputs = submitter.resolve_data(
+                    node_name=node, inputs=inputs, chat_history_name=chat_history_input_name
                 )
-            else:
-                return submitter.flow_test(
-                    inputs=flow_inputs,
-                    environment_variables=environment_variables,
-                    stream_log=stream_log,
-                    stream_output=stream_output,
-                    allow_generator_output=allow_generator_output and is_chat_flow,
-                )
+
+                if node:
+                    return submitter.node_test(
+                        node_name=node,
+                        flow_inputs=flow_inputs,
+                        dependency_nodes_outputs=dependency_nodes_outputs,
+                        environment_variables=environment_variables,
+                        stream=True,
+                    )
+                else:
+                    return submitter.flow_test(
+                        inputs=flow_inputs,
+                        environment_variables=environment_variables,
+                        stream_log=stream_log,
+                        stream_output=stream_output,
+                        allow_generator_output=allow_generator_output and is_chat_flow,
+                    )
 
     @staticmethod
     def _is_chat_flow(flow):
@@ -362,6 +387,8 @@ class FlowOperations(TelemetryMixin):
         node_variant: str = None,
         update_flow_tools_json: bool = True,
     ):
+        # TODO: confirm if we need to import this
+        from promptflow._sdk._submitter import variant_overwrite_context
 
         flow_copy_target = Path(output)
         flow_copy_target.mkdir(parents=True, exist_ok=True)
@@ -556,6 +583,9 @@ class FlowOperations(TelemetryMixin):
 
     @contextlib.contextmanager
     def _resolve_additional_includes(cls, flow_dag_path: Path) -> Iterable[Path]:
+        # TODO: confirm if we need to import this
+        from promptflow._sdk._submitter import remove_additional_includes
+
         if _get_additional_includes(flow_dag_path):
             # Merge the flow folder and additional includes to temp folder.
             # TODO: support a flow_dag_path with a name different from flow.dag.yaml
